extract suggest widget renderer into its own file
